Use Platform.runLater to change GUI
- Use Platform.runLater to change GUI components content
- Add lock to output data lists
- New born predator stay close to its parent
- Change predators and prey fields to float -> more flexible
- Improve Demo 3

+++ src/main/java/CooperativeHunting/Prey.java
@@ -60,7 +60,7 @@ class Prey extends Animal {
      * @param mediumPreyColor:  medium preys' color for visualization
      * @param largePreyColor:   large preys' color for visualization
      */
-    static void set(int speed, float defaultNutrition, int defaultAttack, float visionRadius, int minSize, int maxSize,
+    static void set(int speed, float defaultNutrition, float defaultAttack, float visionRadius, int minSize, int maxSize,
                     Color smallPreyColor, Color mediumPreyColor, Color largePreyColor) {
         Prey.speed = speed;
         Prey.defaultNutrition = defaultNutrition;
